ChatWithCsv/requirements.txt
streamlit
python-dotenv
pandas
langchain
langchain_experimental
langchain_openai
tabulate